Fixed 32-bit warning when smuggling an int through a pointer
git-svn-id: https://svn.code.sf.net/p/vice-emu/code/trunk@40195 379a1393-f5fb-40a0-bcee-ef074d9b53f7

+++ vice/src/arch/gtk3/opengl_renderer.c
@@ -470,7 +470,7 @@ static void vice_opengl_on_ui_frame_clock(GdkFrameClock *clock, video_canvas_t *
 
 static void render(void *job_data, void *pool_data)
 {
-    render_job_t job = (render_job_t)(int)(long long int)job_data;
+    render_job_t job = (render_job_t)(int)(intptr_t)job_data;
     video_canvas_t *canvas = pool_data;
     vice_opengl_renderer_context_t *context = (vice_opengl_renderer_context_t *)canvas->renderer_context;
     backbuffer_t *backbuffer;
